requirements.txt
# Python dependencies for takedownsecondsons
pyyaml>=6.0
requests>=2.28.0
colorama>=0.4.6
python-dotenv>=1.0.0
